Funzioni base per il calcolo di statistiche di legge su dati QA.

stat.period(x, period, necess, FUN = mean)
stat.period2(x, period, nmax.missing, FUN = mean)
which.period(x, period, necess, FUN=which.max)
exc.period(x, period, necess, threshold)

stat.window(x, window, necess, FUN = mean)
mean_window(x, k, necess)

detect.event(x, threshold)

shift(x, k)

Arguments

x

vettore dei valori di concentrazione

period

vettore, della stessa lunghezza di x, che identifica i periodi

window

vettore numerico di due elementi; descrive l'ampiezza della finestra mobile, p.es. c(-7,0) per la media mobile su 8 ore

necess

numero di dati validi necessari in ciascun periodo

nmax.missing

numero massimo di dati mancanti accettabili in ciascun periodo

FUN

funzione da applicare

threshold

soglia

k

in shift, numero di passi di cui si vuole spostare x; in mean_window, ampiezza della finestra

Details

Le funzioni stat.period e stat.period2 calcolano una statistica FUN su periodi definiti, con approcci diversi nella gestione dei mancanti. La funzione which.period lavora come stat.period, ma si puo' usare per funzioni (come which.min o which.max) che non accettano l'argomento na.rm.

Invece stat.window opera su una finestra mobile, e chiama shift che sposta la serie temporale in avanti o indietro nel tempo. Piu' efficiente, ma limitata alla media mobile, e' mean_window.

La funzione exc.period fa il conteggio dei superamenti di una data soglia. Invece detect.event restituisce una matrice contenente informazioni sui superamenti della soglia (vedi sezione Value).

Value

Le funzioni stat.period, stat.period2 e which.period restituiscono un vettore numerico contenente il valore di FUN calcolato su x per ciascuno dei valori di period.

La funzione exc.period restituisce un vettore numerico contenente il conteggio dei superamenti di x per ciascuno dei valori di period.

Le funzioni stat.window e mean_window restituiscono un vettore numerico della stessa lunghezza di x contenenti rispettivamente la statistica stat o la media sulla finestra mobile richiesta.

La funzione detect.event restituisce un data.frame contenente una riga per ciascun evento di superamento, con le colonne

La funzione shift restituisce un vettore numerico della stessa lunghezza di x.