Funzioni per estrarre le elaborazioni sintetiche archiviate sul DB: dbqa.get.elab estrae i valori, dbqa.descr.elab ne visualizza la descrizione.

dbqa.get.elab(con, year, id.param, id.elab,
              type.elab = "F", only.rrqa = T,
              only.valid = T, keep.all = F)
dbqa.descr.elab(con, id.elab=NULL)

Arguments

con

connessione al DB

year

anno

id.param

codice numerico identificativo del parametro

id.elab

codice numerico identificativo dell'elaborazione. In dbqa.descr.elab può essere NULL (è il default) per vedere la denominazione di tutte le elaborazioni definite (molte delle quali non sono però calcolate).

type.elab

tipo di elaborazione:

"I" se numero intero (p.es. numero di superamenti)

"F" se numero reale (p.es. media annua)

only.rrqa

variabile logica, se TRUE tiene solo stazioni RRQA

only.valid

variabile logica, se TRUE tiene solo elaborazioni valide

keep.all

variabile logica, se TRUE estrae tutti i campi della tabella, come sono nel DB

Details

Value