Funzioni per produrre le statistiche annuali tipiche (media annua, superamenti giornalieri della media e del max della media mobile su 8h) per una singola stazione: estrazione, calcoli, scrittura su DB.

prepare.annual_report(con, id.staz, id.param, year=NULL, tstep, ...)
calculate.annual_report(data, id.param, thr.daily.ave=NULL, thr.ave8h.max=NULL,
                        thr.hourly=NULL, thr.multihourly=NULL, NH=3,
                        critical.months=NULL)
write.annual_report(con, AR, id.param, verbose=F, ...)

Arguments

con

identificativo della connessione al DB (stringa)

id.staz

codice numerico identificativo della stazione

id.param

codice numerico identificativo dell'inquinante

year

anno per cui si richiede il report. Se lasciato NULL prende l'anno di 5 mesi fa, facendo riferimento a Sys.Date().

tstep

stringa di carattere che identifica il timestep ("H" o "d", orario o giornaliero)

data

dati estratti da prepare.annual_report

thr.daily.ave

soglia per la media giornaliera

thr.ave8h.max

soglia per il max giornaliero della media mobile su 8h

thr.hourly

soglia oraria

thr.multihourly

soglia per superamenti orari di più ore consecutive

NH

numero di ore consecutive di superamento da contare se thr.multihourly non ? NULL

critical.months

vettore numerico dei mesi su cui calcolare la media di periodo (p.es. c(1:3,10:12) per la media invernale)

AR

lista di data.frame prodotta da calculate.annual_report (vedi Value)

verbose

scrive a video alcune informazioni utili per il debug

…

parametri opzionali. La funzione prepare.annual_report li passa a dbqa.get.datastaz; la funzione write.annual_report li passa a dbqa.insert.

Details

Value

La funzione calculate.annual_report restituisce un data.frame con:

se thr.daily.ave non ? NULL allora data.frame ha anche:

se thr.ave8h.max non ? NULL allora data.frame ha anche:

se thr.hourly o thr.multihourly non sono NULL allora data.frame ha anche:

se thr.hourly non ? NULL allora data.frame ha anche:

se thr.multihourly non ? NULL allora data.frame ha anche:

se critical.months non ? NULL allora data.frame ha anche:

References

Note

See also

Examples

## Not run: ------------------------------------
# db_usr="yourUsername"
# db_pwd="yourPassword"
# db_name="DBaddress"
# con1 <- dbqa.connect(db_usr, db_pwd, db_name)
# dat <- prepare.annual_report(con=con1, id.staz="2000003")
# calculate.annual_report(data=dat)
# dbDisconnect(con1)
## ---------------------------------------------