Funzioni per produrre le statistiche annuali per una singola stazione di ozono: estrazione, calcoli, scrittura su DB.

prepare.ozone_annual_report(con, id.staz, year=NULL, ...)
calculate.ozone_annual_report(data)
write.ozone_annual_report(con, OAR, verbose=F, ...)

Arguments

con

identificativo della connessione al DB (stringa)

id.staz

codice numerico identificativo della stazione

year

anno per cui si richiede il report. Se lasciata NULL prende l'anno di 5 mesi fa, facendo riferimento a Sys.Date().

data

dati estratti da prepare.ozone_annual_report

OAR

lista di data.frame prodotta da calculate.ozone_annual_report (vedi Value)

verbose

scrive a video alcune informazioni utili per il debug

…

parametri opzionali. La funzione prepare.ozone_annual_report li passa a dbqa.get.datastaz; la funzione write.ozone_annual_report li passa a dbqa.insert.

Details

Value

La funzione calculate.ozone_annual_report restituisce una lista di due oggetti:

annual.report è un data.frame di un sola riga con le seguenti colonne:

events è una lista contenente i due data.frame, ciascuno dei quali avente colonne start.time e duration:

References

Note

See also

Examples

## Not run: ------------------------------------
# db_usr="yourUsername"
# db_pwd="yourPassword"
# db_name="DBaddress"
# con1 <- dbqa.connect(db_usr, db_pwd, db_name)
# dat <- prepare.annual_report(con=con1, id.staz="2000003")
# calculate.ozone_annual_report(data=dat)
# dbDisconnect(con1)
## ---------------------------------------------